kthread: set m->holder to next thread when passing mutex
This makes the mutex implementation compatible with non-round-robin
scheduling --- currently, if the current thread schedules before the
next thread to get the lock does, it will think it still holds the mutex
even though it does not (which can cause, among other things, the
reentrancy assertion to fail).

Found this while playing around with a randomized scheduler to try and
make VFS tests more effective at finding race conditions.