vfs: mkdir(): hold lock across creation and get(child)
Without this, there is a chance that the created inode will be deleted
or bound to a different logical entity in between these calls.  Other
creation calls have a similar problem.

The written test doesn't trigger the bug as written --- I wasn't able to
get the thread scheduling to consistently work out in the right way.
With randomized scheduling (rather than round-robin) and forcing threads
to occasionally block in kmutex_unblock() even when the lock is
uncontended, this test consistently triggers a segfault in
vfs_set_created_metadata() as expected.

Including the test under the theory that with future changes and
cranking up test iterations, it will be more effective as a regression
test.