vfs: add open(O_CREAT)/unlink() race test
This tests for a similar race as previously existed in
mkdir/mknod/symlink (except that it's not currently a race in open()).
As with the other tests in this section, this doesn't catch the bug
unless scheduling is randomized, threads sometimes yield during
kthread_unlock(), and iterations are cranked up.  If the bug is added
(and scheduling code modified), it also manifests as a segfault in
vfs_set_created_metadata() due to vfs_get() returning NULL for the
now-deleted child.