vfs: pass removed child vnode to fs unlink/rmdir
This is not functionally necessary, but matches the existing pattern of
passing relevant vnodes, locked, to the filesystem functions.  This also
allows the simultaneous-rename-race test to fail if a changed-source
vnode situation isn't handled correctly.

It also formalizes the contract that the caller will have a reference to
the vnode being unlinked which it will then vfs_put(), at which point
the filesystem will have the opportunity to reclaim/delete the inode.

This could be removed in the future as an optimization to remove the
double lookup.  That would require the filesystem to do more complicated
work, but it is also in a position to do finer-grained locking (for
example, allowing multiple concurrent read operations).