user: add compile-time "header tests" for constants
I discovered while working on login that if you just #include <stdio.h>
from newlib, you get newlib's version of the SEEK_* constants, which are
not the same as the APOS ones.  Which version you get also changes based
on header order (!) --- I've fixed this on the newlib side, but adding
compile-time tests here that include just a single standard header, then
the APOS version, then verify they're set correctly.

The actual assertions don't fire when this fails, currently --- instead
there's an error in the APOS version of the header when it redefines the
constant to a new value.