mmap test: clean up existing mappings at start
This allows the tests to be run in a context where there are already
mappings (e.g. invoked from a user process).  Also, fork into a separate
process to run the tests to preserve whatever mappings there are before
the test was run (this is strictly speaking not necessary, but keeps the
unmapping required for the tests isolated to a separate process so it
can't leak into other tests).