proc: add per-process thread list
This is a functional no-op --- rather than tracking a single thread in
the process_t, we have a list of threads, but we still only ever look at
the first thread on it.

Several tests (FIFO, Unix domain sockets, possibly others) currently
rely on the semi-functional multi-threaded-process support, but they're
well-behaved and the secondary threads always exit before the primary or
process.