signal: handle multiple process threads
Tests will be added later.  Each call site that looks at the process
thread currently is changed to either look at _all_ threads (e.g. signal
delivery), or the _current_ thread (which is asserted to be part of the
current process).

This is a subtle behavior change from today (where the "main" process
thread might be used rather than the current thread).