signal: add unimplemented SIGAPOSTKILL
This (internal) signal will be sent to specific threads to force them to
terminate/exit.  It's internal to the kernel, and shouldn't be allowed
to be triggered by userspace, or sent to a process as a whole.

Adding the (currently unimplemented) SIGACT_TERM_THREAD action, which is
the default/required action for threads receiving the signal.