proc: add proc threads and split exit() into two phases
proc threads are a thin wrapper on top of the kthread primitives.
Currently the only difference is that proc_thread_exit() exits the
current process if the current thread is the last remaining one in the
process.  I'd like to move all process thread management (i.e. adding
and removing the threads from the process thread list) into these
functions as well to improve layering, but that requires updating at
least some tests to use proc_thread_create() rather than
kthread_create() (which is more correct for tests that rely on the
process state, so should be done regardless).

These will eventually be exposed as syscalls, with an additional wrapper
for proc_thread_create() that trampolines into userspace on the supplied
stack and entry address.