user: add futex_ts() syscall
The Linux futex syscall can take either a timespec pointer or another
uint32_t (val2).  For now just adding a wrapper for the timespec version
that handles the 32/64-bit conversion and user-memory-copying in the
wrapper.  When other operations are added that use val2, TBD if we add
another wrapper, move the user memory handling of timespec into the
futex code, or simply expose a syscall/wrapper for each futex operation
directly.