x86: handle double-faults on dedicated stack
Currently a double-fault triggered by e.g. a stack overflow in kernel
mode will cause a triple fault because the double-fault will be
(tried to be) handled on the same bad stack.  This sets up a dedicated
hardware task and TSS to switch to to handle double faults.  To get a
useful backtrace on double faults, the stack unwinder needs to extract
the EIP and EBP from the normal TSS (which show the location the
double-fault occurred).