block cache: handle read errors in BC get/lookup
Any readers who attempt to get the block in question while the operation
is blocking will block until it fails, and then have the error code
propagated back to them.  As soon as the thread that was attempting to
initialize the entry sees the failure, it removes it from the block
cache table, then the last thread holding a reference will clean it up.

This requires some changes to the cleanup flow, and honestly it's sorta
confusing and messy (distinguishing between "real" entries, which are in
the table, and bad entries, which are not in the table).