block cache: keep clearing unpinned until LRU queue is empty
Now, if an entry is added to the LRU queue while
block_cache_clear_unpinned() is running, clearing will continue rather
than crashing the kernel on an assertion failure.

This could theoretically happen before (if another process/thread
decided to put a block while block_cache_clear_unpinned() was running)
but was unlikely.  With the upcoming change to have shadow memobjs keep
pins in their entries, we now can easily get a "cascading release" of
LRU-only (not flushq) entries --- e.g. an entry is released the first
time through the loop, which in turn triggers a memobj to be released,
triggering an unref on a child memobj, which then put()s all of its
pages --- leaving them on the LRU queue (but not the flush queue, since
memobj_shadow uses BC_FLUSH_NONE).