memory: pin all shadow object pages
6+ years after drafting the first fix for this!

This (finally) fixes the bug where a shadow object page could be
destroyed under memory pressure if not in use, even if it could be
referenced later.  This works by adding an additional pin to each shadow
memobj page, which is only released when the object itself is destroyed.
This is triggered by checking if the pin count equals the entry count
(otherwise, there would be a refcount cycle and the memobjs and pages
would never be reclaimed).

The shadow memobj tracks its extant pages in a new hashtable.  This is
somewhat redundant, since the block cache already tracks all the entries
in a list.  However, doing it this way means we don't have to
synchronize accesses and data structure validity between shadow_memobj
and the block cache code, and it's clear which entries have additional
pins and which don't (because they're stored in a dedicated data
structure).

This extends naturally to swapping --- when swapping is implemented,
this code path simply becomes the "swap disabled" support.