memory: add block_cache_free_all()
This will let us proactively flush and deref memobjs that we want to
pull from the block cache.  The motivating use case for this is
unmounting a filesystem --- if a process has mmap'd a file on the
filesystem, it will remain in the block cache until memory pressure
forces it to be removed.

This is a relatively simple operation but makes the current block cache
code structure harder to understand, largely because it adds another
place where entries can be removed/freed (previously was just in the
memory pressue case) --- effectively `flushing` acts as a primitive
refcount on entries to prevent them from being reaped while another
thread is using them (when they aren't pinned).  Other than
block_cache_clear_unpinned(), anytime a blocking operation is done
(which releases the block cache lock), either flushing is set to true,
or the entry pointer is discarded after the operation (because it might
be invalid).  All places that potentially reap a block cache entry check
`flushing` before doing so.

It would likely be better to move to a more obvious refcount system,
even though it is in principle redundant with the current state (the
`flushing` bool, in particular).