kshell: clear block cache before heapprof command
This is almost always what I want to do to make heap profiles
meaningful.  This is easier than running 'test block_cache' every time,
and faster now that the block cache tests take over a second to run.
Also add `hp` abbreviation, and an escape hatch to avoid clearing the
block cache if I want it.