vfs: attempt to flush open vnodes on unmount
Without this, any vnodes that have been mmap'd will likely have pages in
the block cache, which count as references to the vnode and prevent
unmounting.

This approach is inefficient, but should be good enough for current
needs --- rather than attempting to synchronously flush all vnodes until
we find one that can't be flushed down to a refcount of zero, we could
have a more sophisticated way to track references that would let us
simply check whether the filesystem had any non-block-cache references.

In the event of concurrent filesystem activity, there's no guarantee
that the fs's vnode list remains in the same order --- rather than have
a complicated locking structure, we store the vnode number and re-get
the vnode without the central lock held.