block cache: fix cleanup entry handling in free_all
Entries that are deleted (the underlying block is freed) but not fully
cleaned up can still be on the memobj's entry list --- they have to be,
since they still hold a reference on the memobj (that's the point of
async cleanup).  Therefore any entries found via the memobj's entry list
have to be checked to see if they are already cleaned up.

This is another sign that this code is just too complicated at this
point, too much action at a distance.