memory: send SIGBUS on usermode page fault errors
Finally, the small change that required this series of 25+ other
commits.

We still kernel panic on a kernel-mode page fault error --- this should
only happen in the syscall DMZ, but is incorrect (for example, calling
write() and passing the mmap'd read_error address as the source buffer
in the new test will cause a panic).  Future changes will fix that
behavior.