vm: generate SIGSEGVs in vm_resolve_address()
This mirrors the behavior when a page is needed but can't be fetched (it
generates SIGBUS).  The semantics here are slightly weird --- the
pattern from the other checkers here is that this is simply verification
logic --- what to _do_ with that result (e.g. generate a signal) should
be done by the caller.  That pattern was broken when paging was added
here (and thus SIGBUS could be generated), so continuing that here.

An alternative would be for this to propagate up an indication of the
type of fault, and the caller could generate the appropriate signal as
needed.  I think that would be cleaner, but more complicated (today,
certainly).