syscall: generate arch-independent template code
Now rather than generating different code based on the arch (e.g.
dispatching to the 32-bit translator or not), the same code is emitted
for all architectures but the different behavior is enabled/disabled at
compile-time.

This will allow checking in the generated template code without it being
arch-specific.

The generating code (both syscalls.py and the templates) are a bit
convoluted at this point, but cleanup/refactor can happen separately.