syscall: replace memcpy with safe user/kernel copies
The buffer and string copies now use the safe
syscall_copy_{to,from}_user() functions rather than kmemcpy.  This means
that any errors encountered during the copy (such as an IO error when
paging in memory for an mmap'd file) will be handled and translated into
a signal delivered to the userspace process, rather than panicing the
kernel.

This makes calls to syscall_verify_buffer() mostly redundant --- before
removing them, will need to verify NULL-handling (for optional
[NULL-allowed] and non-optional arguments) and address checks.

String (checked with syscall_verify_string()) and pointer table
arguments are still problematic if the memory map is concurrently
changed, or there is an error paging in mmap'd memory while calculating
the string/table length.