build: check in syscall generated .tpl.c files
This makes it easy to see the impact of changes on the generated files,
as diffs in the git history.  Getting the files to be generated in the
source directory rather than the build/variant directory took some
fiddling and experimentation, but this seems to work fine --- the Object
object has to be generated manually to force the .o file to be written
in the build/variant directory, and then it has to be included directly
in the object list rather than the source file (otherwise scons gets
confused about where the file is and how to generate it).