syscall: remove verify_buffer() call in DMZ
Now, rather than verifying each buffer up-front and then again on-demand
as it's copied to/from, we only validate during the copy.  Functionally
the only differences are,
  1) bad buffer arguments now cause signals (SIGSEGV/SIGBUS), rather
     than just returning EFAULT.
  2) for written buffers, the problem will only be detected _after_ the
     syscall (with whatever side effects it includes) is run.  This was
     already the case for 'runtime' errors (generating SIGBUS).

Most of the diff here is updating userspace tests to deal with this new
behavior.

Due to the change, some extra handling is needed for optional buffer
arguments, and to detect bad sizes (e.g. negative numbers).  To get
around weird sign issues, I just cast the size value to size_t and
compare to UINT32_MAX / 2 --- negative values (if the size type is
signed) are therefore rejected, as are too-large values.  I use
UINT32_MAX rather than SIZE_MAX to get the same behavior on 32- and
64-bit builds (for test consistency).

There are certainly more principled ways to do those checks; left a TODO
to cleanup and address in the future.