memory: fix lost unpinned anonymous pages bug
This is the same bug as shadow objects had --- if a page was currently
unpinned in any process, it could be lost.

Fixed by throwing a shadow object in front of all anonymous objects
(since they already handle this case correctly).  This makes it even
sillier to have multiple anonymous objects, much less each with multiple
pages of all zeroes.  Sharing a single page is slightly more complicated
than ideal due to the fact that the offset within the object is keyed as
part of the page, and there's no way for the anonymous memobj to signal
to the block cache that a particular page can be used for multiple
offsets within the object.