memory: fix major copy-on-write/fork bug
Currently if the parent process accesses a page post-fork before the
child, it will see the bc_entry_t in the vm_area's pages list (even
though the area's memobj is now the new shadow object) and simply use
it, writing to shared memory (which is now visible to the child).

The fix here is to simply zero out the pages array and dereference all
pages, in addition to unmapping all the virtual memory.  If we switched
to simply making the virtual memory read-only, then I think the current
logic would be correct (because then any writes would trigger the
fork-page flow).