proc: add mutex deadlock detection
Each thread now tracks all the mutexes it holds.  When a mutex is
locked, we record which other mutexes are being held by the current
thread inside that mutex.  We check on each lock operation that the
current thread is not holding any mutexes that previously were locked
after the new mutex (which would indicate a cycle, and likely deadlock).

With this implementation and settings, test runs are slowed down by a
little under 10%.  I will attempt to optimize a bit; if I can't get that
down, this may need to be disabled by default.