memory: lock object reading subobj for get/read page
This is only localized locking that protects the integrity of the shadow
chain via the subobj field, but not any broader data consistency.
Because the result of get() is used outside the data structure (in the
VM code), potential for weird race conditions here between threads.

We'll need locks in the VM/proc code to ensure consistency of memobj
result usage around calls to get().