block cache: add cache entry migration
This relies on the caller to ensure no other accesses to the entry are
added during the call --- this is needed to ensure that the flushed
state doesn't change during blocking operations (notably getting the
target entry, which could block due to mutex locks or to wait for it to
be initialized).

It's sorta gross that the block cache can't fully guarantee consistency
here, but in practice it's not an issue --- migration is only used by
shadow objects, and they can lock across operations.  Doing it this way
avoid a lot more complexity in the migrate implementation.