memory: implement basic shadow chain collapse
After more than a year in the works :)

Tried doing this many different ways, including on exit() --- after a
lot of work, settled on doing it in fork() for simplicity.  This
defers collapse/migration of shadow objects and their entries until a
fork() call, but is much simpler to implement as it only requires
traversing the shadow chain in a single direction.

On creating a new shadow object, if the child is itself a shadow
object, the chain is collapsed.  In the future, if the chain is too
long, the creation operation will fail (this requires plumbing the
failure back), which will be the ultimate prize.

Currently collapse of the chain is aborted (and the count is incorrect)
if an object fails to be collapsed (e.g. due to a page that is currently
resident in a process even though the memobj itself is not directly
referenced).