memory: use spinlock for shadow object refcount
The shadow object now uses both locks --- a spinlock for the refcount
and cleaning_up bit, and the mutex for the page status and related data
structures.

This is redundant now, but will be necessary later so that the refcount
can be manipulated while the mutex is held, protecting consistency of
the shadow data structures.