memory: add user-test for collapse-over-in-use-page
This currently passes trivially since there is no shadow chain
collapsing or page migration.  It tests the scenario of an
otherwise-unreferenced shadow object in the chain who has a page that is
currently in-use by a process.  In that case, the shodow object must be
retained even though it could otherwise be collapsed.