proc: add guid-based proc wait for tests
This gives tests the ability to atomically wait for a _particular_
process to exit without having to worry about pids being reused.  AFAIK
there's no POSIX equivalent to something like this.

It's not particularly sophisticated, but localizes the logic and data
structures in one place so that tests can be updated down the road (e.g.
if locks are added).