kthread: prioritize enabled threads in mutexes
Currently if a thread is waiting for a mutex and is disabled, it will
take the mutex and be unable to run, blocking everyone else.  Now if
there are any non-disabled threads also waiting, they will get priority
to take the mutex and run.

If only disabled threads are waiting then the first will be selected
(which will still prevent another enabled thread from taking it if it
comes along later).