memory: fix lookup/migrate race on shadow objs
In practice today this is very hard to hit, if not impossible, due to
other locking and how thread scheduling works.  But this fixes the
locking for future-proofness.  Unfortunately the test mucks with shadow
object internals and is pretty brittle due to how it orchestrates thread
scheduling to trigger the race.