memory: add more block cache migration tests
These test for when the migration hits a target entry that is in the
process of being initialized.

Note that this scenario relies on the shadow object ensuring that there
is no concurrent access to the block cache for the object for logical
correctness (the block cache data structures will be valid regardless).