memory: handle unflushed entries in migrate()
We now force-flush any unflushed entries, not just ones that are
actively flushing.  This, along with the requirement that the caller not
let anyone else get the entry during the call, ensures that no one else
(including the flush thread) touches the entry after we check the status
at the beginning.

In turn, this significantly simplies the logic post-getting-target,
since we no longer need to worry about whether the entry in unflushed
(migrating its flushed status), whether it is _now_ actively flushing
(and waiting for the flush thread again), etc.

The cost is that migration is now blocked during flushing, but since
migration is only used by shadow objects and they have no-op flushes
currently this is insignificant.