memory: hold lock for entire shadow obj get_page()
This is not strictly necessary --- for data structure protection, more
fine grained locking in read_page() as previously is sufficient.  The
block cache also requires that no new references to entries are created
while a migrate() call is running.

That's not possible even without this change --- block_cache_lookup()
calls in shadow_get_page() are protected with the lock, and
block_cache_get() is only called when a writable entry is requested.  A
writable entry can only be requested if another thread (other than a
parent object migrating from the object in question) also holds a
reference, in which case we would not be migrating entries.

If we are migrating entries, then the only reference to the child object
is from the parent object.  Another thread _could_ request an entry via
that reference, but only a read-only entry --- which is protected by the
lock around lookup() in get_page().

However, this is simpler to reason about and doesn't require reasoning
across many different call sites.