build/x64: create dedicated non-.text init sections
This parallels 91f9c5b573bf16d08 for i586.  When linking, ld still gives
errors about RWX segments, I believe because the ALIGN(0x1000) forces it
to place .text and non-.text segments next to each other.  Fixing that
will require more work (and it's not clear if it matters, given that
this ELF is not loaded with virtual memory on).