kprintf: add PRIxADDR and more 64-bit tests
This started as a more ambitious change to properly add portable 64-bit
support (with PRIx64) to kprintf.  Adding 64-bit (long long) support to
kprintf itself is actually pretty straightforward, but updating the
underlying kutoa()-etc functions is more annoying due to the lack of
intrinsic 64-bit math support (and defining separate 64-bit versions is
also annoying, for a variety of reasons).

There is not much of a need to portably print 64-bit numbers on all
architectures (vs printing native address sizes, which might be 64 bit),
so just adding a couple tests for now.  Anyone who tries to printf a
64-bit number that's a long long or int64 (or unsigned variant) will get
a compile-time error; if this is important, can add full support later.