riscv64: add basic RISC-V support and kernel stub.
All this does is boot then busy-spin.  The headers define enough for the
rest of the kernel to compile, but the definitions may not be correct
(and of course a lot of functionality is missing).  We can't even
validate or print the FDT yet because ksprintf() calls klog() and
friends.

Booting riscv64 is _much_ simpler than x86/x64 -- the part before
setting up virtual memory is just a few lines of assembly.

The convention for TODOs is TODO(riscv) marks something that needs to be
addressed before merging this into main.  Others may be addressed later.