arch: define interrupts helpers as real functions
These are not being inlined currently anyway, and the boilerplate the
compiler generates is a significant length.  For RISC-V will define
these directly in assembly.

This is some of the oldest code in the kernel, dating back to the first
few weeks.  It also has a bug, needlessly reading the interrupt state on
restore_interrupts() in the x86 version.