riscv: set up kernel physical map in kinit()
We set up a 4GB map for the kernel to use.  This is much larger than
what we do on x86/x64 (256MB), and is also placed below the kernel code
rather than above.  This is to keep the kernel code linked in the last
2GB of the address space (allowing linking with medany to work across
physical and virtual addresses --- a mild convenience).

Since on the virt board the main memory is mapped so far into the
physical address space, we need a much larger mapping.  We will also
need to adjust the page mapping code to account for usable pages not
starting at address 0.