memory: add main memory start to meminfo
This is a no-op on x86/x64 (PC), but will not be on riscv64.  Arguably,
this should be set to the first page after the kernel rather than zero
(and then we can remove that logic from the page allocator).

This really should be a proper memory map system that can supply
multiple discontigous ranges of memory with different properties (such
as is given in the FDT).  That's overkill for now, though.