riscv: share memory layout constants across ld/s/c files
This was a lot more complicated than I anticipated.  I originally
considered using the preprocessor to add constants to the LD and ASM
files but that has various issues (e.g. ASM comments start with '#').

The primary downside of this approach is that we end up with symbols in
the final binary for each constant, multiple times (once from the linker
script, once from the loader.s file).  This doesn't seem to actually
impact anything, but is confusing.

I tried getting the assembly code to use the symbols generated by the
linker, but it doesn't work --- we're relying on the assembler to
generate instructions to create 64-bit literals, so it needs to see the
literals (even if they are just symbols).

This does not changed the compiled executable at all, other than adding
the extraneous symbols.