riscv: fix interrupt handler stack frame
This now creates a normal-looking stack frame that GDB can read.  It
still can't get past the interrupted function if it's a leaf, since ra
isn't saved on the stack and GDB (even with a custom unwinder) can't
find ra.

Also fixes a missing s7 save, and removes a redundant ra save.