build: () -> (void) on function defns
Recent clang versions now warn (error) with -Wstrict-prototypes on
function _definitions_ that are not (void), even when there is an
existing declaration.

In C2x it seems that C will adopt C++ rules that make f() the same as
f(void) --- ideally we'd just adopt that and avoid this churn, but I
don't see an obvious way to do that easily, and this way doesn't
introduce compiler version compatibility issues.