devicetree: add no-alloc tree parsing
This creates a parsed version of the FDT that is much easier to consume,
but still avoids any dynamic memory allocation.  This is the rough moral
equivalent of an FDT if each node included a length.

The extra pointers and alignments (on riscv we align the structs to 8
byte addresses in-memory, not 4 like in the DTB) roughly double the
memory consumed by the raw FDT (~5KB -> ~10KB).

We will need to copy the raw FDT anyway, as it does not reside in
protected memory (at least not with QEMU and OpenSBI).