riscv: call kmain! (and disable legacy PC devices/IOPORTs)
This adds two different knobs (that in practice will always likely be
set to the same value):
 - ARCH_SUPPORTS_IOPORT controls whether the architecture supports
   port-mapped IO.  All architectures are assumed to support MMIO.
 - ARCH_SUPPORTS_LEGACY_PC_DEVS controls whether the architecture is
   assumed to have "legacy" PC devices (such as PS/2, statically-mapped
   IO ports for PCI, VGA, etc).

We skip legacy devices during boot, allowing RISC-V to boot into kmain.
It currently runs off the end because it can't initialize a default TTY
and run the kshell --- but doesn't segfault anymore.

Some of these devices (such as PCI) need to be updated to support
running with flexible MMIO.