ld: support oflag.ONLCR (\n -> \r\n)
This is required when sending to an actual serial device --- previously,
we just "faked" it in vterm.  Rather than send the translation
unilaterally, control properly with ONLCR (default on).

I'm unsure if this should trigger on echoing as well --- if so, maybe it
would be better to put this in ld_term_putc() and call that from
ld_write().

The basic change to vterm handles the typical case (\r\n) correctly ---
more complexity is needed to handle the different combinations on their
own.  Added a bunch more tests for scrolling and wrapping of the various
combinations (\r\n, \r, \n, \n\r).