ld: handle CR -> CRLF on input as well
This consolidates the output-munging code into ld_term_putc(), which now
must be told whether it is echoing or not, since different options apply
(for example ECTOCTL, which is currently always on).  Otherwise
write()ing a control character (e.g. to trigger a color change) gets
escaped.

This is required for portable communication with a remote terminal over
a serial line (tested with telnet on macos, which does not currently
work).