ktest: add KEXPECT_MULTILINE_STREQ()
Hit a fun macro issue here, when I started with typeof(a) for aval and
bval (like KEXPECT_INT_) --- for string literals, the typeof resolves to
an array with a length (e.g. char[2]), so typeof(a) and typeof(b) are
_not_ the same.  GCC compiled passing "\n\n\n" with typeof(a) as char[2]
as a two-word string on the stack without NULL termination.