riscv: push/pop SPP and SPIE in interrupt handler
These control whether we return to user mode and whether interrupts are
re-enabled when executing sret at the end (respectively).  Currently, we
never see a nested interrupt, so they don't change.  With defints,
though, we can see nested interrupts; the processor resets these
unilaterrally to 0 and 1 (user-mode with interrupts enabled) on sret, so
a nested interrupt clobbers the previous values.

Now we save/restore them explicitly on the stack.  This relies on a
nested interrupt not being possible before we save and after we restore
--- so long as the stack is paged in, this should always be true.