devicetree: add basic dtint_map() tracer
This is currently basically a no-op --- without interrupt nexuses and
interrupt maps, the only arrangement to be traced is directly from an
interrupt to its parent, which must be the root --- and dtint_extract()
already does the hard work.

Device drivers can call this as-is, however, and don't have to worry
about the details of the interrupt tree.