riscv: don't assume char is signed in kprintf tests
On riscv, char is signed rather than unsigned (like on x86).  That
breaks the %hhi kprintf tests, because (char)-10 because an _unsigned_
char equal to 246, not a _signed_ char equal to -10.  The bit pattern is
the same, but when passed as a vararg to kprintf, widening doesn't
sign-extend the unsigned char.

Now all tests pass on RISC-V! \o/