memory: make heap profile work with modern pprof
The Go version of pprof is stricter about the legacy header format ---
it's possible this should be heapprofile rather than heap, need to
investigate how sampling works.

The good news is that Go pprof otherwise seems to work great, and unlike
the perl pprof doesn't need modifications to deal with the APOS kernel
binary --- it works out of the box.