riscv: swap stacks on user-mode interrupt
We use two doublewords at the start of the address stored in sscratch
for scratch space to determine if we need to switch stacks.  I think
this could be done with one, or even none, with a more clever approach.

sscratch is defined to point to the bottom (highest address) of the
kernel mode stack buffer for the thread (and the stack itself starts
right after that).  This is guaranteed (barring bugs) to be accessible
even if the kernel has a stack overflow (though it currently doesn't
detect that).  It's not as robust as e.g. double faults on x86 are,
though.