kthread: fix stack length tracking bug
Currently, all stacks are assumed to have a stack of size
KTHREAD_STACK_SIZE, but the initial thread stack may not. On x86 this
never causes a problem because the bottom of the stack is only ever used
on handling an interrupt or trap from userspace, which will never happen
with the initial kthread.

In a future patch, on riscv, the bottom of the stack is used even for
kernel interrupts (in a future change).  Before this change, this
overwrites memory _after_ the kernel stack which happens to be the
initial page table.  This manifests as erasing a page table entry for
the kernel heap, causing us lose a heap page and get another fault the
next time the heap is accessed --- which then pulls in a fresh page,
losing the data that was in the heap and causing shenanigans.