riscv: implement signal handling
This reveals a weird property of the syscall code today, which is that
it assumes that it doesn't get called from an interrupt context (and
therefore has its own context-extraction and return-to-userspace
handling).

It would be better, at the cost of a stack frame, to have an
arch-specific syscall trampoline that handles return-to-userspace,
similar to the arch-specific interrupt handler.  On RISC-V, that
trampoline wouldn't exist.