vfs: readlink() takes bufsize as size_t
This should probably done consistently for the rest of these functions,
but readlink is the only one right now that doesn't match its syscall
definition, so starting there.  Enforce that the passed in size can be
represented by a signed int before passing to the FS code (which still
uses int).