syscall: limit size of buffers to 256 pages
Allocating large buffers for syscalls is costly, and requires allocating
enough physical pages to fill the entire buffer passed.  If a large
buffer size is supplied, this could even exhaust physical memory just
allocating (and copying in/out) the kernel buffer for the syscall.

This sets an arbitrary limit of 256 pages on syscall buffers.  In the
future, this could be configurable.  Practially speaking, this is only
relevant in riscv64 user tests that pass large buffers --- on x86/64,
the buffers passed are large enough to cause kmalloc() to fail to find a
buffer, not to fail to allocate pages (which causes a kernel panic).