vga: only call legacy IO functions when legacy VGA exists
In practice the VGA code isn't called on non-VGA platforms anyway (i.e.
riscv64), other than from the vterm tests.  We could just disable vterm
tests on riscv64, but that logic isn't strongly tied to VGA or the
underlying platform, so it's nicer to continue to be able to run/test
it.

It would be cleaner to make these virtual function calls against an
abstract "video" interface, but IIRC this is a hotspot --- and this code
isn't really safe or optimal anyway, so why put lipstick on the pig.